// src/common/enums/module-type.enum.ts
export enum ModuleTypeEnum {
  SINTER_BLEND = '烧结配料计算',
  BF_BLEND_SINGLE = '单独高炉配料计算',
  IRON_PRE_INTEGRATION_I = '铁前一体化配料计算I',
  IRON_PRE_INTEGRATION_II = '铁前一体化配料计算II',
  PROFIT_INTEGRATION = '利润一体化配料计算',
}


export enum ModuleName {
  SJ_BL = '烧结配料计算',
  BF_BL = '单独高炉配料计算',
  IRON_I = '铁前一体化配料计算I',
  IRON_II = '铁前一体化配料计算II',
  PROFIT = '利润一体化配料计算',
  SJ_ECON = '烧结原料经济性评价',
  COKE_ECON = '焦炭经济性评价',
  PELLET_ECON = '外购球团块矿经济性评价',
  LUMP_ECON = '外购块矿经济性评价',
  PCI_ECON = '喷吹煤经济性评价',
  MIX_COAL = '混合煤性价比计算',
  COST_PROFIT = '现价成本与利润测算',
  SJ_FIXED = '烧结固定配料计算',
  S_BALANCE = '硫平衡计算',
  PIG_FIXED = '生铁固定配料计算',
}